+++ modules/hooks/hooks.js
@@ -1,14 +1,15 @@
 import { MODULE_ID, SYSTEM_SOCKET } from '../tables/constants.js';
 import { HMActiveEffectHooks } from './effect-hooks.js';
 import { HMSupport } from '../sys/support.js';
-import { HMActor } from '../actor/actor.js';
+import { HMActorHooks } from './actor-hooks.js';
+import { HMChatHooks } from './chat-hooks.js';
+import { HMItemHooks } from './item-hooks.js';
+import { HMTokenHooks } from './token-hooks.js';
 import { HMWeaponItem } from '../item/weapon-item.js';
-import { HMChatMgr } from '../mgr/chatmgr.js';
 import { HMCombat, HMCombatTracker } from '../sys/combat.js';
 import { HMToken } from '../sys/token.js';
 import { HMMacro } from '../sys/macro.js';
 import { handleSocketEvent } from '../sys/sockets.js';
-import { HMItemHooks } from './item-hooks.js';
 
 async function ready() {
     if (game.modules.get('_dev-mode')?.api?.getPackageDebugValue(MODULE_ID)) {
@@ -32,13 +33,13 @@ export const registerHooks = () => {
     Hooks.on('applyActiveEffect', HMActiveEffectHooks.applyActiveEffect);
     Hooks.on('createActiveEffect', HMActiveEffectHooks.createActiveEffect);
     Hooks.on('deleteActiveEffect', HMActiveEffectHooks.deleteActiveEffect);
-    Hooks.on('createActor', HMActor.createActor);
-    Hooks.on('createToken', HMActor.createToken);
+    Hooks.on('createActor', HMActorHooks.createActor);
+    Hooks.on('createToken', HMTokenHooks.createToken);
     Hooks.on('hoverToken', (token, state) => token.drawReach(state));
     Hooks.on('createItem', HMItemHooks.createItem);
     Hooks.on('preCreateItem', HMItemHooks.preCreateItem);
     Hooks.on('updateItem', HMWeaponItem.updateItem);
-    Hooks.on('renderChatMessage', HMChatMgr.renderChatMessage);
+    Hooks.on('renderChatMessage', HMChatHooks.renderChatMessage);
     Hooks.on('updateCombat', HMCombat.updateCombat);
     Hooks.on('createCombatant', HMCombat.createCombatant);
     Hooks.on('deleteCombatant', HMCombat.deleteCombatant);