+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+- Cleaned up ranged attack chat cards.
+
 ## [0.3.5] - 2022-12-18
 ### Added
 - Sort Spell Library by 'all'.