+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.3.6] - 2022-12-22
+### Added
+- Compendium items.
 ### Fixed
 - Cleaned up ranged attack chat cards.
 - Cleaned up beast ToP notification.